Arybo word-level expressions and lazy eval
Add a word-level expression AST inside arybo, which allows to define
expressions like ExprXor(ExprBV(mba.var('x')), ExprCst(0xAA)).
We can also eval them bit-per-bit, which makes a kind of "lazy
evaluation" for these expressions. For instance, with the expression
ExprAdd(x32,y32)[:8] (only the first 8 bits of the 32-bit addition of x
and y), we will only compute the first 8 bits of this addition (and not
the 32 bits of the addition and then keep only the first 8 bits).